tests/wpt/tests/html/canvas/element/path-objects/2d.path.closed.arc.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.path.closed.arc</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.path.closed.arc</h1>
<p class="desc">line caps should not be drawn on closed arcs</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="300" height="300"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("line caps should not be drawn on closed arcs");
_addTest(function(canvas, ctx) {

  ctx.beginPath();
  ctx.strokeStyle = 'red';
  ctx.lineWidth = 20;
  ctx.lineCap = 'square';
  ctx.arc(150, 150, 50, 0, 2 * Math.PI);
  ctx.closePath();
  ctx.stroke();
  // The pixel (209, 148) is part of the line cap, but not part of the circle.
  // Linecap is not drawn if the circle is closed, so the pixel is not red.
  assert_not_equals(ctx.getImageData(209, 148, 1, 1).data, [255, 0, 0, 255])

});
</script>
